First commit pre changes

+++ WindowsFormsApp1/MeasureModel.cs
@@ -9,6 +9,7 @@ namespace WindowsFormsApp1
     public class MeasureModel
     {
         public System.DateTime DateTime { get; set; }
-        public double Value { get; set; }
+        public float Value { get; set; }
+      
     }
 }
